+++ assets/js/dashboard/stats/graph/visitor-graph.js
@@ -5,7 +5,7 @@ import { navigateToQuery } from '../../query'
 import * as api from '../../api'
 import * as storage from '../../util/storage'
 import LazyLoader from '../../components/lazy-loader'
-import {GraphTooltip, buildDataSet, METRIC_MAPPING, METRIC_LABELS, METRIC_FORMATTER} from './graph-util';
+import {GraphTooltip, buildDataSet, METRIC_MAPPING, METRIC_LABELS, METRIC_FORMATTER, LoadingState} from './graph-util';
 import dateFormatter from './date-formatter';
 import TopStats from './top-stats';
 import { IntervalPicker, getStoredInterval, storeInterval } from './interval-picker';
@@ -13,10 +13,16 @@ import FadeIn from '../../fade-in';
 import * as url from '../../util/url'
 import classNames from "classnames";
 
-const LOADING_STATE = {
-  loading: 'loading',
-  refreshing: 'refreshing',
-  loaded: 'loaded'
+const calculateMaximumY = function(dataset) {
+  const yAxisValues = dataset
+    .flatMap((item) => item.data)
+    .map((item) => item || 0)
+
+  if (yAxisValues) {
+    return Math.max(...yAxisValues)
+  } else {
+    return 1
+  }
 }
 
 class LineGraph extends React.Component {
@@ -34,9 +40,7 @@ class LineGraph extends React.Component {
     const { graphData, metric, query } = this.props
     const graphEl = document.getElementById("main-graph-canvas")
     this.ctx = graphEl.getContext('2d');
-    const dataSet = buildDataSet(graphData.plot, graphData.present_index, this.ctx, METRIC_LABELS[metric])
-    // const prev_dataSet = graphData.prev_plot && buildDataSet(graphData.prev_plot, false, this.ctx, METRIC_LABELS[metric], true)
-    // const combinedDataSets = comparison.enabled && prev_dataSet ? [...dataSet, ...prev_dataSet] : dataSet;
+    const dataSet = buildDataSet(graphData.plot, graphData.comparison_plot, graphData.present_index, this.ctx, METRIC_LABELS[metric])
 
     return new Chart(this.ctx, {
       type: 'line',
@@ -62,7 +66,8 @@ class LineGraph extends React.Component {
         onClick: this.onClick.bind(this),
         scales: {
           y: {
-            beginAtZero: true,
+            min: 0,
+            suggestedMax: calculateMaximumY(dataSet),
             ticks: {
               callback: METRIC_FORMATTER[metric],
               maxTicksLimit: 8,
@@ -72,6 +77,12 @@ class LineGraph extends React.Component {
               zeroLineColor: 'transparent',
               drawBorder: false,
             }
+          },
+          yComparison: {
+            min: 0,
+            suggestedMax: calculateMaximumY(dataSet),
+            display: false,
+            grid: { display: false },
           },
           x: {
             grid: { display: false },
@@ -147,14 +158,14 @@ class LineGraph extends React.Component {
   }
 
   componentDidMount() {
-    if (this.props.metric && this.props.graphData) {
+    if (this.props.graphData) {
       this.chart = this.regenerateChart();
     }
     window.addEventListener('mousemove', this.repositionTooltip);
   }
 
   componentDidUpdate(prevProps) {
-    const { graphData, metric, darkTheme } = this.props;
+    const { graphData, darkTheme } = this.props;
     const tooltip = document.getElementById('chartjs-tooltip');
 
     if (
@@ -162,7 +173,7 @@ class LineGraph extends React.Component {
       darkTheme !== prevProps.darkTheme
     ) {
 
-      if (metric && graphData) {
+      if (graphData) {
         if (this.chart) {
           this.chart.destroy();
         }
@@ -175,7 +186,7 @@ class LineGraph extends React.Component {
       }
     }
 
-    if (!graphData || !metric) {
+    if (!graphData) {
       if (this.chart) {
         this.chart.destroy();
       }
@@ -337,8 +348,8 @@ export default class VisitorGraph extends React.Component {
   constructor(props) {
     super(props)
     this.state = {
-      topStatsLoadingState: LOADING_STATE.loading,
-      mainGraphLoadingState: LOADING_STATE.loading,
+      topStatsLoadingState: LoadingState.loading,
+      mainGraphLoadingState: LoadingState.loading,
       metric: storage.getItem(`metric__${this.props.site.domain}`) || 'visitors'
     }
     this.onVisible = this.onVisible.bind(this)
@@ -369,12 +380,12 @@ export default class VisitorGraph extends React.Component {
   updateInterval(interval) {
     if (this.isIntervalValid(interval)) {
       storeInterval(this.props.query.period, this.props.site.domain, interval)
-      this.setState({ mainGraphLoadingState: LOADING_STATE.refreshing }, this.fetchGraphData)
+      this.setState({ mainGraphLoadingState: LoadingState.refreshing }, this.fetchGraphData)
     }
   }
 
   onVisible() {
-    this.setState({mainGraphLoadingState: LOADING_STATE.loading}, this.fetchGraphData)
+    this.setState({mainGraphLoadingState: LoadingState.loading}, this.fetchGraphData)
     this.fetchTopStatData()
     if (this.props.query.period === 'realtime') {
       document.addEventListener('tick', this.fetchGraphData)
@@ -383,38 +394,34 @@ export default class VisitorGraph extends React.Component {
   }
 
   componentDidUpdate(prevProps, prevState) {
-    const { metric, topStatData } = this.state;
-    const { query, site } = this.props
+    const { metric } = this.state;
+    const { query } = this.props
 
     if (query !== prevProps.query) {
-      if (this.isGraphCollapsed()) {
-        this.setState({ topStatsLoadingState: LOADING_STATE.loading, topStatData: null })
-      } else {
-        this.setState({ mainGraphLoadingState: LOADING_STATE.loading, topStatsLoadingState: LOADING_STATE.loading, graphData: null, topStatData: null }, this.fetchGraphData)
-      }
+      this.setState({ mainGraphLoadingState: LoadingState.loading, topStatsLoadingState: LoadingState.loading, graphData: null, topStatData: null }, this.fetchGraphData)
       this.fetchTopStatData()
     }
 
     if (metric !== prevState.metric) {
-      this.setState({mainGraphLoadingState: LOADING_STATE.refreshing}, this.fetchGraphData)
-    }
-
-    const savedMetric = storage.getItem(`metric__${site.domain}`)
-    const topStatLabels = topStatData && topStatData.top_stats.map(({ name }) => METRIC_MAPPING[name]).filter(name => name)
-    const prevTopStatLabels = prevState.topStatData && prevState.topStatData.top_stats.map(({ name }) => METRIC_MAPPING[name]).filter(name => name)
-    if (topStatLabels && `${topStatLabels}` !== `${prevTopStatLabels}`) {
-      if (query.filters.goal && metric !== 'conversions') {
-        this.setState({ metric: 'conversions' })
-      } else if (topStatLabels.includes(savedMetric) && savedMetric !== "") {
-        this.setState({ metric: savedMetric })
-      } else {
-        this.setState({ metric: topStatLabels[0] })
-      }
+      this.setState({mainGraphLoadingState: LoadingState.refreshing}, this.fetchGraphData)
     }
   }
 
-  isGraphCollapsed() {
-    return this.state.metric === ""
+  resetMetric() {
+    const { topStatData } = this.state
+    const { query, site } = this.props
+
+    const savedMetric = storage.getItem(`metric__${site.domain}`)
+    const selectableMetrics = topStatData && topStatData.top_stats.map(({ name }) => METRIC_MAPPING[name]).filter(name => name)
+    const canSelectSavedMetric = selectableMetrics && selectableMetrics.includes(savedMetric)
+
+    if (query.filters.goal) {
+      this.setState({ metric: 'conversions' })
+    } else if (canSelectSavedMetric) {
+      this.setState({ metric: savedMetric })
+    } else {
+      this.setState({ metric: 'visitors' })
+    }
   }
 
   componentWillUnmount() {
@@ -423,18 +430,13 @@ export default class VisitorGraph extends React.Component {
   }
 
   updateMetric(clickedMetric) {
-    const newMetric = clickedMetric === this.state.metric ? "" : clickedMetric
+    if (this.state.metric == clickedMetric) return
 
-    storage.setItem(`metric__${this.props.site.domain}`, newMetric)
-    this.setState({ metric: newMetric })
+    storage.setItem(`metric__${this.props.site.domain}`, clickedMetric)
+    this.setState({ metric: clickedMetric, graphData: null })
   }
 
   fetchGraphData() {
-    if (this.isGraphCollapsed()) {
-      this.setState({ mainGraphLoadingState: LOADING_STATE.loaded, graphData: null })
-      return
-    }
-
     const url = `/api/stats/${encodeURIComponent(this.props.site.domain)}/main-graph`
     let params = { metric: this.state.metric }
     const interval = this.getIntervalFromStorage()
@@ -442,19 +444,19 @@ export default class VisitorGraph extends React.Component {
 
     api.get(url, this.props.query, params)
       .then((res) => {
-        this.setState({ mainGraphLoadingState: LOADING_STATE.loaded, graphData: res })
+        this.setState({ mainGraphLoadingState: LoadingState.loaded, graphData: res })
         return res
       })
       .catch((err) => {
         console.log(err)
-        this.setState({ mainGraphLoadingState: LOADING_STATE.loaded, graphData: false })
+        this.setState({ mainGraphLoadingState: LoadingState.loaded, graphData: false })
       })
   }
 
   fetchTopStatData() {
     api.get(`/api/stats/${encodeURIComponent(this.props.site.domain)}/top-stats`, this.props.query)
       .then((res) => {
-        this.setState({ topStatsLoadingState: LOADING_STATE.loaded, topStatData: res })
+        this.setState({ topStatsLoadingState: LoadingState.loaded, topStatData: res }, this.resetMetric)
         return res
       })
   }
@@ -465,15 +467,13 @@ export default class VisitorGraph extends React.Component {
 
     const theme = document.querySelector('html').classList.contains('dark') || false
 
-    const topStatsLoadedOrRefreshing = (topStatsLoadingState === LOADING_STATE.loaded || topStatsLoadingState === LOADING_STATE.refreshing)
-    const mainGraphLoadedOrRefreshing = (mainGraphLoadingState === LOADING_STATE.loaded || mainGraphLoadingState === LOADING_STATE.refreshing)
-    const noMetricOrRefreshing = (!metric || mainGraphLoadingState === LOADING_STATE.refreshing)
+    const mainGraphRefreshing = (mainGraphLoadingState === LoadingState.refreshing)
     const topStatAndGraphLoaded = !!(topStatData && graphData)
 
     const showGraph =
-    topStatsLoadedOrRefreshing &&
-    mainGraphLoadedOrRefreshing &&
-      (topStatData && noMetricOrRefreshing || topStatAndGraphLoaded)
+      LoadingState.isLoadedOrRefreshing(topStatsLoadingState) &&
+      LoadingState.isLoadedOrRefreshing(mainGraphLoadingState) &&
+      (topStatData && mainGraphRefreshing || topStatAndGraphLoaded)
 
     return (
       <FadeIn show={showGraph}>
@@ -483,23 +483,20 @@ export default class VisitorGraph extends React.Component {
   }
 
   render() {
-    const {metric, mainGraphLoadingState, topStatsLoadingState} = this.state
+    const {mainGraphLoadingState, topStatsLoadingState} = this.state
     const loaderClassName = classNames('mx-auto loading', {
-      'pt-52 sm:pt-56 md:pt-60': mainGraphLoadingState == LOADING_STATE.refreshing,
-      'pt-32 sm:pt-36 md:pt-48': mainGraphLoadingState !== LOADING_STATE.refreshing && metric,
-      'pt-16 sm:pt-14 md:pt-18 lg:pt-5': mainGraphLoadingState !== LOADING_STATE.refreshing && !metric
+      'pt-52 sm:pt-56 md:pt-60': mainGraphLoadingState == LoadingState.refreshing,
+      'pt-32 sm:pt-36 md:pt-48': mainGraphLoadingState !== LoadingState.refreshing,
     })
 
-    const loadingOrRefreshing =
-          mainGraphLoadingState == LOADING_STATE.refreshing ||
-          mainGraphLoadingState == LOADING_STATE.loading ||
-          topStatsLoadingState == LOADING_STATE.refreshing ||
-          topStatsLoadingState == LOADING_STATE.loading
+    const showLoader =
+      LoadingState.isLoadingOrRefreshing(mainGraphLoadingState) ||
+      LoadingState.isLoadingOrRefreshing(topStatsLoadingState)
 
     return (
       <LazyLoader onVisible={this.onVisible}>
-        <div className={`relative w-full mt-2 bg-white rounded shadow-xl dark:bg-gray-825 transition-padding ease-in-out duration-150 ${metric ? 'main-graph' : 'top-stats-only'}`}>
-          {loadingOrRefreshing && <div className="graph-inner"><div className={loaderClassName}><div></div></div></div>}
+        <div className={"relative w-full mt-2 bg-white rounded shadow-xl dark:bg-gray-825 main-graph"}>
+          {showLoader && <div className="graph-inner"><div className={loaderClassName}><div></div></div></div>}
           {this.renderInner()}
         </div>
       </LazyLoader>