+++ lib/plausible/application.ex
@@ -9,6 +9,7 @@ defmodule Plausible.Application do
     children = [
       Plausible.Repo,
       Plausible.ClickhouseRepo,
+      Plausible.IngestRepo,
       {Finch, name: Plausible.Finch, pools: finch_pool_config()},
       {Phoenix.PubSub, name: Plausible.PubSub},
       Plausible.Session.Salts,
@@ -60,7 +61,7 @@ defmodule Plausible.Application do
   end
 
   defp maybe_add_sentry_pool(pool_config) do
-    case Application.get_env(:sentry, :dsn) do
+    case Sentry.Config.dsn() do
       dsn when is_binary(dsn) ->
         Map.put(pool_config, dsn, size: 50)