+++ lib/plausible/stats/base.ex
@@ -6,6 +6,7 @@ defmodule Plausible.Stats.Base do
   # Ecto typespec has not been updated for this PR: https://github.com/elixir-ecto/ecto/pull/3592
   @dialyzer {:nowarn_function, add_sample_hint: 2}
   @no_ref "Direct / None"
+  @not_set "(not set)"
 
   def base_event_query(site, query) do
     events_q = query_events(site, query)
@@ -14,7 +15,9 @@ defmodule Plausible.Stats.Base do
       sessions_q =
         from(
           s in query_sessions(site, query),
-          select: %{session_id: s.session_id}
+          select: %{session_id: s.session_id},
+          where: s.sign == 1,
+          group_by: s.session_id
         )
 
       from(
@@ -389,6 +392,7 @@ defmodule Plausible.Stats.Base do
   defp db_prop_val(:utm_campaign, @no_ref), do: ""
   defp db_prop_val(:utm_content, @no_ref), do: ""
   defp db_prop_val(:utm_term, @no_ref), do: ""
+  defp db_prop_val(_, @not_set), do: ""
   defp db_prop_val(_, val), do: val
 
   def utc_boundaries(%Query{period: "realtime"}, _timezone) do